refactor: generalize ProgressIndicator into InfoBox for loading and error messages
